Struct syntex_syntax::ast::Mac_

pub struct Mac_ {
    pub path: Path,
    pub tts: Vec<TokenTree>,
    pub ctxt: SyntaxContext,
}

Represents a macro invocation. The Path indicates which macro is being invoked, and the vector of token-trees contains the source of the macro invocation.

NB: the additional ident for a macro_rules-style macro is actually stored in the enclosing item. Oog.

Fields

path

tts

ctxt