Struct syntex_syntax::ast::Path

pub struct Path {
    pub span: Span,
    pub global: bool,
    pub segments: Vec<PathSegment>,
}

A "Path" is essentially Rust's notion of a name; for instance: std::cmp::PartialEq . It's represented as a sequence of identifiers, along with a bunch of supporting information.

Fields
span: Span global: bool

A ::foo path, is relative to the crate root rather than current module (like paths in an import).

segments: Vec<PathSegment>

The segments in the path: the things separated by ::.