Struct syntex_syntax::ast::PathSegment

pub struct PathSegment {
    pub identifier: Ident,
    pub span: Span,
    pub parameters: Option<P<PathParameters>>,
}

A segment of a path: an identifier, an optional lifetime, and a set of types.

E.g. std, String or Box<T>

Fields
identifier: Ident

The identifier portion of this path segment.

span: Span

Span of the segment identifier.

parameters: Option<P<PathParameters>>

Type/lifetime parameters attached to this path. They come in two flavors: Path<A,B,C> and Path(A,B) -> C. Note that this is more than just simple syntactic sugar; the use of parens affects the region binding rules, so we preserve the distinction. The Option<P<..>> wrapper is purely a size optimization; None is used to represent both Path and Path<>.

Methods

impl PathSegment

pub fn from_ident(ident: Ident, span: Span) -> Self

pub fn crate_root() -> Self