Struct syntex_syntax::codemap::ExpnInfo

pub struct ExpnInfo {
    pub call_site: Span,
    pub callee: NameAndSpan,
}

Extra information for tracking spans of macro and syntax sugar expansion

Fields
call_site: Span

The location of the actual macro invocation or syntax sugar , e.g. let x = foo!(); or if let Some(y) = x {}

This may recursively refer to other macro invocations, e.g. if foo!() invoked bar!() internally, and there was an expression inside bar!; the call_site of the expression in the expansion would point to the bar! invocation; that call_site span would have its own ExpnInfo, with the call_site pointing to the foo! invocation.

callee: NameAndSpan

Information about the expansion.