Struct syntex_syntax::codemap::NameAndSpan

pub struct NameAndSpan {
    pub format: ExpnFormat,
    pub allow_internal_unstable: bool,
    pub span: Option<Span>,
}

Fields

format

The format with which the macro was invoked.

allow_internal_unstable

Whether the macro is allowed to use #[unstable]/feature-gated features internally without forcing the whole crate to opt-in to them.

span

The span of the macro definition itself. The macro may not have a sensible definition span (e.g. something defined completely inside libsyntax) in which case this is None.

Methods

impl NameAndSpan

fn name(&self) -> Name