Struct syntex_syntax::codemap::Span

pub struct Span {
    pub lo: BytePos,
    pub hi: BytePos,
    pub expn_id: ExpnId,
}

Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

Fields

lo

hi

expn_id

Information about where the macro came from, if this piece of code was created by a macro expansion.

Methods

impl Span

fn substitute_dummy(self, other: Span) -> Span

Returns self if self is not the dummy span, and other otherwise.

fn contains(self, other: Span) -> bool