Module syntex_syntax::diagnostic

Reexports

pub use self::Level::*;

pub use self::RenderSpan::*;

pub use self::ColorConfig::*;

Structs

EmitterWriter

ExplicitBug
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

SpanHandler
A span-handler is like a handler but also accepts span information for source-location reporting.

Enums

ColorConfig

Level

RenderSpan

Traits

Emitter

Functions

expect