Struct syntex_syntax::diagnostic::SpanHandler

pub struct SpanHandler {
    pub handler: Handler,
    pub cm: CodeMap,
}

A span-handler is like a handler but also accepts span information for source-location reporting.

Fields

handler

cm

Methods

impl SpanHandler

fn new(handler: Handler, cm: CodeMap) -> SpanHandler

fn span_fatal(&self, sp: Span, msg: &str) -> FatalError

fn span_fatal_with_code(&self, sp: Span, msg: &str, code: &str) -> FatalError

fn span_err(&self, sp: Span, msg: &str)

fn span_err_with_code(&self, sp: Span, msg: &str, code: &str)

fn span_warn(&self, sp: Span, msg: &str)

fn span_warn_with_code(&self, sp: Span, msg: &str, code: &str)

fn span_note(&self, sp: Span, msg: &str)

fn span_end_note(&self, sp: Span, msg: &str)

fn span_help(&self, sp: Span, msg: &str)

fn span_suggestion(&self, sp: Span, msg: &str, suggestion: String)

Prints out a message with a suggested edit of the code.

See diagnostic::RenderSpan::Suggestion for more information.

fn fileline_note(&self, sp: Span, msg: &str)

fn fileline_help(&self, sp: Span, msg: &str)

fn span_bug(&self, sp: Span, msg: &str) -> !

fn span_bug_no_panic(&self, sp: Span, msg: &str)

fn span_unimpl(&self, sp: Span, msg: &str) -> !

fn handler<'a>(&'a self) -> &'a Handler