Module syntex_syntax::diagnostics::metadata

This module contains utilities for outputting metadata for diagnostic errors.

Each set of errors is mapped to a metadata file by a name, which is currently always a crate name.

Structs

ErrorLocation
JSON encodable error location type with filename and line number.

ErrorMetadata
JSON encodable/decodable version of ErrorInfo.

Functions

get_metadata_dir
Get the directory where metadata for a given prefix should be stored.

output_metadata
Write metadata for the errors in err_map to disk, to a file corresponding to prefix/name.

Type Definitions

ErrorMetadataMap
Mapping from error codes to metadata that can be (de)serialized.