Function syntex_syntax::diagnostics::metadata::output_metadata

pub fn output_metadata(ecx: &ExtCtxt, prefix: &str, name: &str, err_map: &ErrorMap) -> Result<(), Box<Error>>

Write metadata for the errors in err_map to disk, to a file corresponding to prefix/name.

For our current purposes the prefix is the target architecture and the name is a crate name. If an error occurs steps will be taken to ensure that no file is created.