Module syntex_syntax::errors

Reexports

pub use errors::emitter::ColorConfig;

Modules

emitter

Structs

ExplicitBug
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

Enums

Level

RenderSpan

Functions

expect