Struct syntex_syntax::ext::base::ExtCtxt

pub struct ExtCtxt<'a> {
    pub parse_sess: &'a ParseSess,
    pub ecfg: ExpansionConfig<'a>,
    pub crate_root: Option<&'static str>,
    pub resolver: &'a mut Resolver,
    pub resolve_err_count: usize,
    pub current_expansion: ExpansionData,
    pub expansions: HashMap<Span, Vec<String>>,
}

One of these is made during expansion and incrementally updated as we go; when a macro expansion occurs, the resulting nodes have the backtrace() -> expn_info of their expansion context stored into their span.

Fields
parse_sess: &'a ParseSess ecfg: ExpansionConfig<'a> crate_root: Option<&'static str> resolver: &'a mut Resolver resolve_err_count: usize current_expansion: ExpansionData expansions: HashMap<Span, Vec<String>>
Methods

impl<'a> ExtCtxt<'a>

pub fn new(
parse_sess: &'a ParseSess,
ecfg: ExpansionConfig<'a>,
resolver: &'a mut Resolver
) -> ExtCtxt<'a>

pub fn expander<'b>(&'b mut self) -> MacroExpander<'b, 'a>

Returns a Folder for deeply expanding all macros in an AST node.

pub fn monotonic_expander<'b>(&'b mut self) -> MacroExpander<'b, 'a>

Returns a Folder that deeply expands all macros and assigns all node ids in an AST node. Once node ids are assigned, the node may not be expanded, removed, or otherwise modified.

pub fn new_parser_from_tts(&self, tts: &[TokenTree]) -> Parser<'a>

pub fn codemap(&self) -> &'a CodeMap

pub fn parse_sess(&self) -> &'a ParseSess

pub fn cfg(&self) -> &CrateConfig

pub fn call_site(&self) -> Span

pub fn backtrace(&self) -> SyntaxContext

pub fn expansion_cause(&self) -> Option<Span>

Returns span for the macro which originally caused the current expansion to happen.

Stops backtracing at include! boundary.

pub fn struct_span_warn(&self, sp: Span, msg: &str) -> DiagnosticBuilder<'a>

pub fn struct_span_err(&self, sp: Span, msg: &str) -> DiagnosticBuilder<'a>

pub fn struct_span_fatal(&self, sp: Span, msg: &str) -> DiagnosticBuilder<'a>

pub fn span_fatal(&self, sp: Span, msg: &str) -> !

Emit msg attached to sp, and stop compilation immediately.

span_err should be strongly preferred where-ever possible: this should only be used when

continuing has a high risk of flow-on errors (e.g. errors in declaring a macro would cause all uses of that macro to complain about "undefined macro"), or

there is literally nothing else that can be done (however, in most cases one can construct a dummy expression/item to substitute; we never hit resolve/type-checking so the dummy value doesn't have to match anything)

pub fn span_err(&self, sp: Span, msg: &str)

Emit msg attached to sp, without immediately stopping compilation.

Compilation will be stopped in the near future (at the end of the macro expansion phase).

pub fn span_warn(&self, sp: Span, msg: &str)

pub fn span_unimpl(&self, sp: Span, msg: &str) -> !

pub fn span_bug(&self, sp: Span, msg: &str) -> !

pub fn trace_macros_diag(&self)

pub fn bug(&self, msg: &str) -> !

pub fn trace_macros(&self) -> bool

pub fn set_trace_macros(&mut self, x: bool)

pub fn ident_of(&self, st: &str) -> Ident

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn std_path(&self, components: &[&str]) -> Vec<Ident>

pub fn name_of(&self, st: &str) -> Name

pub fn check_unused_macros(&self)