Module syntex_syntax::ext::deriving

The compiler code necessary to implement the #[derive] extensions.

FIXME (#2810): hygiene. Search for "__" strings (in other files too). We also assume "extra" is the standard library, and "std" is the core library.

Modules

bounds

clone

debug

decodable
The compiler code necessary for #[derive(Decodable)]. See encodable.rs for more.

default

encodable
The compiler code necessary to implement the #[derive(Encodable)] (and Decodable, in decodable.rs) extension. The idea here is that type-defining items may be tagged with #[derive(Encodable, Decodable)].

eq

generic
Some code that abstracts away much of the boilerplate of writing derive instances for traits. Among other things it manages getting access to the fields of the 4 different sorts of structs and enum variants, as well as creating the method and impl ast instances.

hash

ord

partial_eq

partial_ord

primitive

Functions

register_all