Enum syntex_syntax::ext::deriving::generic::SubstructureFields

pub enum SubstructureFields<'a> {
    Struct(Vec<FieldInfo<'a>>),
    EnumMatching(usize, &'a Variant, Vec<FieldInfo<'a>>),
    EnumNonMatchingCollapsed(Vec<Ident>, &'a [P<Variant>], &'a [Ident]),
    StaticStruct(&'a VariantData, StaticFields),
    StaticEnum(&'a EnumDef, Vec<(Ident, Span, StaticFields)>),
}

A summary of the possible sets of fields.

Variants

Struct

EnumMatching

Matching variants of the enum: variant index, ast::Variant, fields: the field name is only non-None in the case of a struct variant.

EnumNonMatchingCollapsed

Non-matching variants of the enum, but with all state hidden from the consequent code. The first component holds Idents for all of the Self arguments; the second component is a slice of all of the variants for the enum itself, and the third component is a list of Idents bound to the variant index values for each of the actual input Self arguments.

StaticStruct

A static method where Self is a struct.

StaticEnum

A static method where Self is an enum.