Function syntex_syntax::ext::deriving::generic::cs_same_method

pub fn cs_same_method<F>(f: F, enum_nonmatch_f: EnumNonMatchCollapsedFunc, cx: &mut ExtCtxt, trait_span: Span, substructure: &Substructure) -> P<Expr> where F: FnOnce(&mut ExtCtxt, Span, Vec<P<Expr>>) -> P<Expr>

Call the method that is being derived on all the fields, and then process the collected results. i.e.

f(cx, span, vec![self_1.method(__arg_1_1, __arg_2_1),
                 self_2.method(__arg_1_2, __arg_2_2)])