Module syntex_syntax::ext::deriving::generic::ty

A mini version of ast::Ty, which is easier to use, and features an explicit Self type to use when specifying impls to be derived.

Reexports

pub use self::PtrTy::*;

pub use self::Ty::*;

Structs

LifetimeBounds
Lifetimes and bounds on type parameters

Path
A path, e.g. ::std::option::Option::<i32> (global). Has support for type parameters and a lifetime.

Enums

PtrTy
The types of pointers

Ty
A type. Supports pointers, Self, and literals

Functions

borrowed

borrowed_explicit_self

borrowed_ptrty

borrowed_self

get_explicit_self

nil_ty