Struct syntex_syntax::ext::expand::PatIdentRenamer

pub struct PatIdentRenamer<'a> {
    // some fields omitted
}

A tree-folder that applies every rename in its list to the idents that are in PatIdent patterns. This is more narrowly focused than IdentRenamer, and is needed for FnDecl, where we want to rename the args but not the fn name or the generics etc.