Struct syntex_syntax::ext::proc_macro_shim::prelude::Span

pub struct Span {
    pub lo: BytePos,
    pub hi: BytePos,
    pub expn_id: ExpnId,
}

Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

Fields
lo: BytePos hi: BytePos expn_id: ExpnId

Information about where the macro came from, if this piece of code was created by a macro expansion.

Methods

impl Span

fn end_point(self) -> Span

Returns a new span representing just the end-point of this span

fn substitute_dummy(self, other: Span) -> Span

Returns self if self is not the dummy span, and other otherwise.

fn contains(self, other: Span) -> bool

fn source_equal(&self, other: &Span) -> bool

Return true if the spans are equal with regards to the source text.

Use this instead of == when either span could be generated code, and you only care that they point to the same bytes of source text.

fn merge(self, other: Span) -> Option<Span>

Returns Some(span), a union of self and other, on overlap.

fn trim_start(self, other: Span) -> Option<Span>

Returns Some(span), where the start is trimmed by the end of other