Module syntex_syntax::ext::quote::rt

Quasiquoting works via token trees.

This is registered as a set of expression syntax extension called quote! that lifts its argument token-tree to an AST representing the construction of the same token tree, with token::SubstNt interpreted as antiquotes (splices).

Re-exports

pub use parse::new_parser_from_tts;

pub use codemap::dummy_spanned;

Structs

BytePos
A byte offset. Keep this small (currently 32-bits), as AST contains a lot of them.

Span
Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source. You must be careful if the span crosses more than one file - you will not be able to use many of the functions on spans in codemap and you cannot assume that the length of the span = hi - lo; there may be space in the BytePos range between files.

Constants

DUMMY_SP

Traits

ExtParseUtils

ToTokens