Module syntex_syntax::ext::source_util

Functions

expand_column

expand_file
file!(): expands to the current filename */ The filemap (loc.file) contains a bunch more information we could spit out if we wanted.

expand_include
include! : parse the given file as an expr This is generally a bad idea because it's going to behave unhygienically.

expand_include_bytes

expand_include_str

expand_line
line!(): expands to the current line number

expand_mod

expand_stringify