Enum syntex_syntax::ext::tt::macro_parser::ParseResult

pub enum ParseResult<T> {
    Success(T),
    Failure(Span, Token),
    Error(Span, String),
}

Variants
Success(T)Failure(Span, Token)

Arm failed to match. If the second parameter is token::Eof, it indicates an unexpected end of macro invocation. Otherwise, it indicates that no rules expected the given token.

Error(Span, String)

Fatal error (malformed macro?). Abort compilation.