Function syntex_syntax::ext::tt::transcribe::new_tt_reader_with_doc_flag

pub fn new_tt_reader_with_doc_flag<'a>(sp_diag: &'a Handler, interp: Option<HashMap<Name, Rc<NamedMatch>>>, imported_from: Option<Ident>, src: Vec<TokenTree>, desugar_doc_comments: bool) -> TtReader<'a>

The extra desugar_doc_comments flag enables reading doc comments like any other attribute which consists of meta and surrounding #[ ] tokens.

This can do Macro-By-Example transcription. On the other hand, if src contains no TokenTree::Sequences, MatchNts or SubstNts, interp can (and should) be None.