Struct syntex_syntax::json::DiagnosticSpanLine

struct DiagnosticSpanLine {
    text: String,
    highlight_start: usize,
    highlight_end: usize,
}

Fields
text: String highlight_start: usize

1-based, character offset in self.text.

highlight_end: usize
Methods

impl DiagnosticSpanLine

fn line_from_filemap(fm: &FileMap,
index: usize,
h_start: usize,
h_end: usize)
-> DiagnosticSpanLine

fn from_span(span: Span, je: &JsonEmitter) -> Vec<DiagnosticSpanLine>

Create a list of DiagnosticSpanLines from span - each line with any part of span gets a DiagnosticSpanLine, with the highlight indicating the span within the line.