Module syntex_syntax::parse::classify

Routines the parser uses to classify AST nodes

Functions

expr_is_simple_block

expr_requires_semi_to_be_stmt
Does this expression require a semicolon to be treated as a statement? The negation of this: 'can this expression be used as a statement without a semicolon' -- is used as an early-bail-out in the parser so that, for instance, if true {...} else {...} |x| 5 isn't parsed as (if true {...} else {...} | x) | 5

stmt_ends_with_semi
this statement requires a semicolon after it. note that in one case (stmt_semi), we've already seen the semicolon, and thus don't need another.