Module syntex_syntax::parse::lexer::comments

Reexports

pub use self::CommentStyle::*;

Structs

Comment

Literal

Enums

CommentStyle

Functions

all_whitespace
Returns None if the first col chars of s contain a non-whitespace char. Otherwise returns Some(k) where k is first char offset after that leading whitespace. Note k may be outside bounds of s.

consume_comment

consume_whitespace_counting_blank_lines

doc_comment_style

gather_comments_and_literals

is_doc_comment

push_blank_line_comment

read_block_comment

read_line_comments

read_shebang_comment

strip_doc_comment_decoration

trim_whitespace_prefix_and_push_line