Struct syntex_syntax::parse::lexer::StringReader

pub struct StringReader<'a> {
    pub span_diagnostic: &'a Handler,
    pub next_pos: BytePos,
    pub pos: BytePos,
    pub col: CharPos,
    pub ch: Option<char>,
    pub filemap: Rc<FileMap>,
    pub terminator: Option<BytePos>,
    pub save_new_lines: bool,
    pub peek_tok: Token,
    pub peek_span: Span,
    pub fatal_errs: Vec<DiagnosticBuilder<'a>>,
    source_text: Rc<String>,
}

Fields
span_diagnostic: &'a Handler next_pos: BytePos

The absolute offset within the codemap of the next character to read

pos: BytePos

The absolute offset within the codemap of the current character

col: CharPos

The column of the next character to read

ch: Option<char>

The current character (which has been read from self.pos)

filemap: Rc<FileMap> terminator: Option<BytePos>

If Some, stop reading the source at this position (inclusive).

save_new_lines: bool

Whether to record new-lines in filemap. This is only necessary the first time a filemap is lexed. If part of a filemap is being re-lexed, this should be set to false.

peek_tok: Token peek_span: Span fatal_errs: Vec<DiagnosticBuilder<'a>> source_text: Rc<String>
Methods

impl<'a> StringReader<'a>

fn new_raw<'b>(span_diagnostic: &'b Handler,
filemap: Rc<FileMap>)
-> StringReader<'b>

For comments.rs, which hackily pokes into next_pos and ch

fn new_raw_internal<'b>(span_diagnostic: &'b Handler,
filemap: Rc<FileMap>)
-> StringReader<'b>

fn new<'b>(span_diagnostic: &'b Handler,
filemap: Rc<FileMap>)
-> StringReader<'b>

fn ch_is(&self, c: char) -> bool

fn fatal_span(&self, sp: Span, m: &str) -> FatalError

Report a fatal lexical error with a given span.

fn err_span(&self, sp: Span, m: &str)

Report a lexical error with a given span.

fn fatal_span_(&self, from_pos: BytePos, to_pos: BytePos, m: &str) -> FatalError

Report a fatal error spanning [from_pos, to_pos).

fn err_span_(&self, from_pos: BytePos, to_pos: BytePos, m: &str)

Report a lexical error spanning [from_pos, to_pos).

fn fatal_span_char(&self,
from_pos: BytePos,
to_pos: BytePos,
m: &str,
c: char)
-> FatalError

Report a lexical error spanning [from_pos, to_pos), appending an escaped character to the error message

fn struct_fatal_span_char(&self,
from_pos: BytePos,
to_pos: BytePos,
m: &str,
c: char)
-> DiagnosticBuilder<'a>

fn err_span_char(&self, from_pos: BytePos, to_pos: BytePos, m: &str, c: char)

Report a lexical error spanning [from_pos, to_pos), appending an escaped character to the error message

fn struct_err_span_char(&self,
from_pos: BytePos,
to_pos: BytePos,
m: &str,
c: char)
-> DiagnosticBuilder<'a>

fn fatal_span_verbose(&self,
from_pos: BytePos,
to_pos: BytePos,
m: String)
-> FatalError

Report a lexical error spanning [from_pos, to_pos), appending the offending string to the error message

fn advance_token(&mut self) -> Result<(), ()>

Advance peek_tok and peek_span to refer to the next token, and possibly update the interner.

fn byte_offset(&self, pos: BytePos) -> BytePos

fn with_str_from<T, F>(&self, start: BytePos, f: F) -> T where F: FnOnce(&str) -> T

Calls f with a string slice of the source text spanning from start up to but excluding self.pos, meaning the slice does not include the character self.ch.

fn name_from(&self, start: BytePos) -> Name

Create a Name from a given offset to the current offset, each adjusted 1 towards each other (assumes that on either side there is a single-byte delimiter).

fn name_from_to(&self, start: BytePos, end: BytePos) -> Name

As name_from, with an explicit endpoint.

fn with_str_from_to<T, F>(&self, start: BytePos, end: BytePos, f: F) -> T where F: FnOnce(&str) -> T

Calls f with a string slice of the source text spanning from start up to but excluding end.

fn translate_crlf<'b>(&self,
start: BytePos,
s: &'b str,
errmsg: &'b str)
-> Cow<'b, str>

Converts CRLF to LF in the given string, raising an error on bare CR.

fn bump(&mut self)

Advance the StringReader by one character. If a newline is discovered, add it to the FileMap's list of line start offsets.

fn nextch(&self) -> Option<char>

fn nextch_is(&self, c: char) -> bool

fn nextnextch(&self) -> Option<char>

fn nextnextch_is(&self, c: char) -> bool

fn scan_optional_raw_name(&mut self) -> Option<Name>

Eats *, if possible.

fn scan_comment(&mut self) -> Option<TokenAndSpan>

PRECONDITION: self.ch is not whitespace Eats any kind of comment.

fn scan_whitespace_or_comment(&mut self) -> Option<TokenAndSpan>

If there is whitespace, shebang, or a comment, scan it. Otherwise, return None.

fn scan_block_comment(&mut self) -> Option<TokenAndSpan>

Might return a sugared-doc-attr

fn scan_digits(&mut self, real_radix: u32, scan_radix: u32) -> usize

Scan through any digits (base scan_radix) or underscores, and return how many digits there were.

real_radix represents the true radix of the number we're interested in, and errors will be emitted for any digits between real_radix and scan_radix.

fn scan_number(&mut self, c: char) -> Lit

Lex a LIT_INTEGER or a LIT_FLOAT

fn scan_hex_digits(&mut self,
n_digits: usize,
delim: char,
below_0x7f_only: bool)
-> bool

Scan over n_digits hex digits, stopping at delim, reporting an error if too many or too few digits are encountered.

fn scan_char_or_byte(&mut self,
start: BytePos,
first_source_char: char,
ascii_only: bool,
delim: char)
-> bool

Scan for a single (possibly escaped) byte or char in a byte, (non-raw) byte string, char, or (non-raw) string literal. start is the position of first_source_char, which is already consumed.

Returns true if there was a valid char/byte, false otherwise.

fn scan_unicode_escape(&mut self, delim: char) -> bool

Scan over a \u{...} escape

At this point, we have already seen the \ and the u, the { is the current character. We will read at least one digit, and up to 6, and pass over the }.

fn scan_float_exponent(&mut self)

Scan over a float exponent.

fn check_float_base(&mut self,
start_bpos: BytePos,
last_bpos: BytePos,
base: usize)

Check that a base is valid for a floating literal, emitting a nice error if it isn't.

fn binop(&mut self, op: BinOpToken) -> Token

fn next_token_inner(&mut self) -> Result<Token, ()>

Return the next token from the string, advances the input past that token, and updates the interner

fn consume_whitespace(&mut self)

fn read_to_eol(&mut self) -> String

fn read_one_line_comment(&mut self) -> String

fn consume_non_eol_whitespace(&mut self)

fn peeking_at_comment(&self) -> bool

fn scan_byte(&mut self) -> Lit

fn scan_byte_escape(&mut self, delim: char, below_0x7f_only: bool) -> bool

fn scan_byte_string(&mut self) -> Lit

fn scan_raw_byte_string(&mut self) -> Lit