Module syntex_syntax::parse::parser

Reexports

pub use self::PathParsingMode::*;

Structs

ModulePath
Information about the path to a module.

ModulePathError

ModulePathSuccess

Parser

Restrictions

Enums

BoundParsingMode
How to parse a bound, whether to allow bound modifiers such as ?.

LhsExpr

ParsePub
pub should be parsed in struct fields and not parsed in variant fields

PathParsingMode
How to parse a path. There are four different kinds of paths, all of which are parsed somewhat differently.

TokenType

Constants

RESTRICTION_NO_STRUCT_LITERAL

RESTRICTION_STMT_EXPR