Module syntex_syntax::parse::token

Reexports

pub use self::BinOpToken::*;

pub use self::Nonterminal::*;

pub use self::DelimToken::*;

pub use self::IdentStyle::*;

pub use self::Lit::*;

pub use self::Token::*;

Modules

keywords
All the valid words that have meaning in the Rust language.

special_idents

special_names

Structs

InternedString
Represents a string stored in the thread-local interner. Because the interner lives for the life of the thread, this can be safely treated as an immortal string, as long as it never crosses between threads.

Enums

BinOpToken

DelimToken
A delimiter token

IdentStyle

Lit

Nonterminal
For interpolation during macro expansion.

SpecialMacroVar

Token

Constants

SELF_KEYWORD_NAME

SELF_KEYWORD_NAME_NUM

Functions

fresh_mark

fresh_name

gensym
gensym's a new usize, using the current interner.

gensym_ident
Maps a string to a gensym'ed identifier.

get_ident_interner

intern
Maps a string to its interned representation.

intern_and_get_ident
Interns and returns the string contents of an identifier, using the thread-local interner.

reset_ident_interner
Reset the ident interner to its initial state.

str_to_ident
Maps a string to an identifier with an empty syntax context.

Type Definitions

IdentInterner