Enum syntex_syntax::parse::token::Token

pub enum Token {
    Eq,
    Lt,
    Le,
    EqEq,
    Ne,
    Ge,
    Gt,
    AndAnd,
    OrOr,
    Not,
    Tilde,
    BinOp(BinOpToken),
    BinOpEq(BinOpToken),
    At,
    Dot,
    DotDot,
    DotDotDot,
    Comma,
    Semi,
    Colon,
    ModSep,
    RArrow,
    LArrow,
    FatArrow,
    Pound,
    Dollar,
    Question,
    OpenDelim(DelimToken),
    CloseDelim(DelimToken),
    Literal(Lit, Option<Name>),
    Ident(Ident),
    Underscore,
    Lifetime(Ident),
    Interpolated(Rc<Nonterminal>),
    DocComment(Name),
    SubstNt(Ident),
    Whitespace,
    Comment,
    Shebang(Name),
    Eof,
}

Variants
EqLtLeEqEqNeGeGtAndAndOrOrNotTildeBinOp(BinOpToken)BinOpEq(BinOpToken)AtDotDotDotDotDotDotCommaSemiColonModSepRArrowLArrowFatArrowPoundDollarQuestionOpenDelim(DelimToken)

An opening delimiter, eg. {

CloseDelim(DelimToken)

A closing delimiter, eg. }

Literal(Lit, Option<Name>)Ident(Ident)UnderscoreLifetime(Ident)Interpolated(Rc<Nonterminal>)DocComment(Name)

Doc comment

SubstNt(Ident)

A syntactic variable that will be filled in by macro expansion.

Whitespace

Whitespace

Comment

Comment

Shebang(Name)Eof
Methods

impl Token

pub fn is_like_gt(&self) -> bool

Returns true if the token starts with '>'.

pub fn can_begin_expr(&self) -> bool

Returns true if the token can appear at the start of an expression.

pub fn can_begin_type(&self) -> bool

Returns true if the token can appear at the start of a type.

pub fn is_lit(&self) -> bool

Returns true if the token is any literal

pub fn ident(&self) -> Option<Ident>

pub fn is_ident(&self) -> bool

Returns true if the token is an identifier.

pub fn is_doc_comment(&self) -> bool

Returns true if the token is a documentation comment.

pub fn is_interpolated(&self) -> bool

Returns true if the token is interpolated.

pub fn is_path(&self) -> bool

Returns true if the token is an interpolated path.

pub fn is_lifetime(&self) -> bool

Returns true if the token is a lifetime.

pub fn is_mutability(&self) -> bool

Returns true if the token is either the mut or const keyword.

pub fn is_qpath_start(&self) -> bool

pub fn is_path_start(&self) -> bool

pub fn is_keyword(&self, kw: Keyword) -> bool

Returns true if the token is a given keyword, kw.

pub fn is_path_segment_keyword(&self) -> bool

pub fn is_any_keyword(&self) -> bool

Returns true if the token is either a strict or reserved keyword.

pub fn is_strict_keyword(&self) -> bool

Returns true if the token is a strict keyword.

pub fn is_reserved_keyword(&self) -> bool

Returns true if the token is a keyword reserved for possible future use.