Module syntex_syntax::print::pprust

Reexports

pub use self::AnnNode::*;

Structs

CurrentCommentAndLiteral

NoAnn

State

Enums

AnnNode

Constants

DEFAULT_COLUMNS

INDENT_UNIT

Traits

PpAnn

PrintState

Functions

arg_to_string

arm_to_string

attr_to_string

attribute_to_string

binop_to_string

block_to_string

bounds_to_string

explicit_self_to_string

expr_to_string

fn_block_to_string

fun_to_string

generics_to_string

ident_to_string

impl_item_to_string

item_to_string

lifetime_to_string

lit_to_string

mac_to_string

meta_item_to_string

pat_to_string

path_to_string

print_crate
Requires you to pass an input filename and reader so that it can scan the input text for comments and literals to copy forward.

rust_printer

rust_printer_annotated

stmt_to_string

to_string

token_to_string

trait_item_to_string

tt_to_string

tts_to_string

ty_param_to_string

ty_to_string

variant_to_string

visibility_qualified

where_clause_to_string