Function syntex_syntax::print::pprust::print_crate

pub fn print_crate<'a>(cm: &'a CodeMap, span_diagnostic: &Handler, krate: &Crate, filename: String, input: &mut Read, out: Box<Write + 'a>, ann: &'a PpAnn, is_expanded: bool) -> Result<()>

Requires you to pass an input filename and reader so that it can scan the input text for comments and literals to copy forward.