Module syntex_syntax::test

Structs

EntryPointCleaner

Test

TestCtxt

TestHarnessGenerator

Enums

HasTestSignature

ShouldPanic

Functions

generate_test_harness

ignored_span
Craft a span that will be ignored by the stability lint's call to codemap's is_internal check. The expanded code calls some unstable functions in the test crate.

is_bench_fn

is_ignored

is_test_crate

is_test_fn

mk_main

mk_reexport_mod

mk_std

mk_test_desc_and_fn_rec

mk_test_descs

mk_test_module

mk_tests

modify_for_testing

nospan

path_name_i

path_node

should_panic