Struct syntex_syntax::tokenstream::TokenStream

pub struct TokenStream { /* fields omitted */ }

Token Streams

A TokenStream is an abstract sequence of tokens, organized into TokenTrees. The goal is for procedural macros to work with TokenStreams and TokenTrees instead of a representation of the abstract syntax tree. Today's TokenTrees can still contain AST via Token::Interpolated for back-compat.

Methods

impl TokenStream

pub fn empty() -> TokenStream

pub fn is_empty(&self) -> bool

pub fn concat(streams: Vec<TokenStream>) -> TokenStream

Important traits for Cursor
impl Iterator for Cursor type Item = TokenTree;

pub fn trees(&self) -> Cursor

Important traits for Cursor
impl Iterator for Cursor type Item = TokenTree;

pub fn into_trees(self) -> Cursor

pub fn eq_unspanned(&self, other: &TokenStream) -> bool

Compares two TokenStreams, checking equality without regarding span information.