Function syntex_syntax::util::lev_distance::find_best_match_for_name

pub fn find_best_match_for_name<'a, T>(iter_names: T, lookup: &str, dist: Option<usize>) -> Option<InternedString> where T: Iterator<Item=&'a Name>

To find the best match for a given string from an iterator of names As a loose rule to avoid the obviously incorrect suggestions, it takes an optional limit for the maximum allowable edit distance, which defaults to one-third of the given word