Module syntex_syntax::ast_util

Structs

IdRange

IdRangeComputingVisitor

IdVisitor
A visitor that applies its operation to all of the node IDs in a visitable thing.

Traits

IdVisitingOperation

Functions

compute_id_range_for_fn_body
Computes the id range for a single fn body, ignoring nested items.

ident_to_pat

ident_to_path

impl_pretty_name
Generate a "pretty" name for an impl from its type and trait. This is designed so that symbols of impl'd methods give some hint of where they came from, (previously they would all just be listed as __extensions__::method_name::hash, with no indication of the type).

is_path

pat_is_ident
Returns true if the given pattern consists solely of an identifier and false otherwise.

path_name_eq

path_name_i

path_to_ident

segments_name_eq

struct_field_visibility