Struct syntex_syntax::parse::lexer::StringReader

pub struct StringReader<'a> {
    pub sess: &'a ParseSess,
    pub next_pos: BytePos,
    pub pos: BytePos,
    pub col: CharPos,
    pub ch: Option<char>,
    pub filemap: Rc<FileMap>,
    pub terminator: Option<BytePos>,
    pub save_new_lines_and_multibyte: bool,
    pub peek_tok: Token,
    pub peek_span: Span,
    pub fatal_errs: Vec<DiagnosticBuilder<'a>>,
    // some fields omitted
}

Fields
sess: &'a ParseSess next_pos: BytePos

The absolute offset within the codemap of the next character to read

pos: BytePos

The absolute offset within the codemap of the current character

col: CharPos

The column of the next character to read

ch: Option<char>

The current character (which has been read from self.pos)

filemap: Rc<FileMap> terminator: Option<BytePos>

If Some, stop reading the source at this position (inclusive).

save_new_lines_and_multibyte: bool

Whether to record new-lines and multibyte chars in filemap. This is only necessary the first time a filemap is lexed. If part of a filemap is being re-lexed, this should be set to false.

peek_tok: Token peek_span: Span fatal_errs: Vec<DiagnosticBuilder<'a>>
Methods

impl<'a> StringReader<'a>

pub fn parse_all_token_trees(&mut self) -> PResult<'a, TokenStream>

impl<'a> StringReader<'a>

pub fn real_token(&mut self) -> TokenAndSpan

pub fn try_next_token(&mut self) -> Result<TokenAndSpan, ()>

Return the next token. EFFECT: advances the string_reader.

pub fn emit_fatal_errors(&mut self)

pub fn peek(&self) -> TokenAndSpan

impl<'a> StringReader<'a>

pub fn new_raw(sess: &'a ParseSess, filemap: Rc<FileMap>) -> Self

For comments.rs, which hackily pokes into next_pos and ch

pub fn new(sess: &'a ParseSess, filemap: Rc<FileMap>) -> Self

pub fn retokenize(sess: &'a ParseSess, span: Span) -> Self

pub fn ch_is(&self, c: char) -> bool

pub fn fatal_span(&self, sp: Span, m: &str) -> FatalError

Report a fatal lexical error with a given span.

pub fn err_span(&self, sp: Span, m: &str)

Report a lexical error with a given span.

pub fn with_str_from<T, F>(&self, start: BytePos, f: F) -> T where
F: FnOnce(&str) -> T,

Calls f with a string slice of the source text spanning from start up to but excluding self.pos, meaning the slice does not include the character self.ch.

pub fn name_from(&self, start: BytePos) -> Name

Create a Name from a given offset to the current offset, each adjusted 1 towards each other (assumes that on either side there is a single-byte delimiter).

pub fn name_from_to(&self, start: BytePos, end: BytePos) -> Name

As name_from, with an explicit endpoint.

pub fn bump(&mut self)

Advance the StringReader by one character. If a newline is discovered, add it to the FileMap's list of line start offsets.

pub fn nextch(&self) -> Option<char>

pub fn nextch_is(&self, c: char) -> bool

pub fn nextnextch(&self) -> Option<char>

pub fn nextnextch_is(&self, c: char) -> bool