Enum syntex_syntax::parse::token::Token

pub enum Token {
    Eq,
    Lt,
    Le,
    EqEq,
    Ne,
    Ge,
    Gt,
    AndAnd,
    OrOr,
    Not,
    Tilde,
    BinOp(BinOpToken),
    BinOpEq(BinOpToken),
    At,
    Dot,
    DotDot,
    DotDotDot,
    Comma,
    Semi,
    Colon,
    ModSep,
    RArrow,
    LArrow,
    FatArrow,
    Pound,
    Dollar,
    Question,
    OpenDelim(DelimToken),
    CloseDelim(DelimToken),
    Literal(Lit, Option<Name>),
    Ident(Ident, IdentStyle),
    Underscore,
    Lifetime(Ident),
    Interpolated(Nonterminal),
    DocComment(Name),
    MatchNt(Ident, Ident, IdentStyle, IdentStyle),
    SubstNt(Ident, IdentStyle),
    SpecialVarNt(SpecialMacroVar),
    Whitespace,
    Comment,
    Shebang(Name),
    Eof,
}

Variants

Eq

Lt

Le

EqEq

Ne

Ge

Gt

AndAnd

OrOr

Not

Tilde

BinOp

BinOpEq

At

Dot

DotDot

DotDotDot

Comma

Semi

Colon

ModSep

RArrow

LArrow

FatArrow

Pound

Dollar

Question

OpenDelim

An opening delimiter, eg. {

CloseDelim

A closing delimiter, eg. }

Literal

Ident

Underscore

Lifetime

Interpolated

DocComment

Doc comment

MatchNt

Parse a nonterminal (name to bind, name of NT, styles of their idents)

SubstNt

A syntactic variable that will be filled in by macro expansion.

SpecialVarNt

A macro variable with special meaning.

Whitespace

Whitespace

Comment

Comment

Shebang

Eof

Methods

impl Token

fn is_like_gt(&self) -> bool

Returns true if the token starts with '>'.

fn can_begin_expr(&self) -> bool

Returns true if the token can appear at the start of an expression.

fn is_lit(&self) -> bool

Returns true if the token is any literal

fn is_ident(&self) -> bool

Returns true if the token is an identifier.

fn is_path(&self) -> bool

Returns true if the token is an interpolated path.

fn is_plain_ident(&self) -> bool

Returns true if the token is a path that is not followed by a :: token.

fn is_lifetime(&self) -> bool

Returns true if the token is a lifetime.

fn is_mutability(&self) -> bool

Returns true if the token is either the mut or const keyword.

fn to_binop(&self) -> Option<BinOp_>

Maps a token to its corresponding binary operator.

fn is_keyword(&self, kw: Keyword) -> bool

Returns true if the token is a given keyword, kw.

fn is_keyword_allow_following_colon(&self, kw: Keyword) -> bool

fn is_any_keyword(&self) -> bool

Returns true if the token is either a special identifier, or a strict or reserved keyword.

fn is_strict_keyword(&self) -> bool

Returns true if the token may not appear as an identifier.

fn is_reserved_keyword(&self) -> bool

Returns true if the token is a keyword that has been reserved for possible future use.

fn mtwt_eq(&self, other: &Token) -> bool

Hygienic identifier equality comparison.

See styntax::ext::mtwt.